requirements.txt
aiohttp==3.6.2
amqp==5.0.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
billiard==3.6.3.0
cachetools==4.1.1
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.1
google-api-core==1.22.2
google-auth==1.22.0
google-cloud-core==1.4.1
google-cloud-translate==3.0.1
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
importlib-metadata==2.0.0
kombu==5.0.2
libcst==0.3.11
multidict==4.7.6
mypy-extensions==0.4.3
oauthlib==3.1.0
prompt-toolkit==3.0.7
proto-plus==1.10.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
tweepy==3.9.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.0
zipp==3.2.0
